Perbaikan data master buku

+++ app/Controllers/Admin/MasterBuku/Buku.php
@@ -1,20 +1,30 @@
 <?php
 
-namespace App\Controllers\Admin;
+namespace App\Controllers\Admin\MasterBuku;
 
 use App\Controllers\BaseController;
-use App\Models\AnggotaModel;
 use App\Models\BukuModel;
+use App\Models\KategoriModel;
+use App\Models\PenerbitModel;
+use App\Models\PenulisModel;
+use App\Models\RakModel;
 
 class Buku extends BaseController
 {
-    protected $bukuModel;
-    protected $anggotaModel;
+    protected $buku;
+    protected $penulis;
+    protected $kategori;
+    protected $penerbit;
+    protected $rak;
 
     public function __construct()
     {
-        $this->bukuModel = new BukuModel();
-        $this->anggotaModel = new AnggotaModel();
+        $this->buku = new BukuModel();
+        $this->penulis = new PenulisModel();
+        $this->penerbit = new PenerbitModel();
+        $this->kategori = new KategoriModel();
+        $this->rak = new RakModel();
+        $this->data += ["inNavBuku" => true];
     }
 
     public function index()
@@ -24,14 +34,18 @@ class Buku extends BaseController
             "subtitle" => "Buku",
             "navactive" => "buku",
             "validation" => validation_errors(),
-            "data" => $this->bukuModel->ambilBuku(),
+            "data" => $this->buku->getDataBySlug(),
+            "dataKategori" => $this->kategori->findAll(),
+            "dataPenulis" => $this->penulis->findAll(),
+            "dataPenerbit" => $this->penerbit->findAll(),
+            "dataRak" => $this->rak->findAll(),
         ];
         return view('admin/buku/dataBuku', $this->data);
     }
 
     public function detail($slug)
     {
-        $buku = $this->bukuModel->ambilBuku($slug);
+        $buku = $this->buku->getDataBySlug($slug);
         if (empty($buku)) {
             throw new \CodeIgniter\Exceptions\PageNotFoundException('Buku "' . $slug . '" tidak ditemukan');
         }
@@ -40,7 +54,11 @@ class Buku extends BaseController
             "title" => "Buku | " .  $buku['judul'],
             "data" => $buku,
             "navactive" => "buku",
-            "validation" => validation_errors()
+            "validation" => validation_errors(),
+            "dataKategori" => $this->kategori->findAll(),
+            "dataPenulis" => $this->penulis->findAll(),
+            "dataPenerbit" => $this->penerbit->findAll(),
+            "dataRak" => $this->rak->findAll(),
         ];
 
         return view('admin/buku/detailBuku', $this->data);
@@ -53,7 +71,6 @@ class Buku extends BaseController
 
     public function simpan()
     {
-
         if (!$this->validate([
             'judul' => [
                 'rules' => 'required|is_unique[buku.judul]',
@@ -63,15 +80,31 @@ class Buku extends BaseController
                 ]
             ],
             'penulis' => [
-                'rules' => 'required',
+                'rules' => 'required|is_not_unique[penulis.id_penulis]',
                 'errors' => [
-                    'required' => 'Kolom penulis tidak boleh kosong'
+                    'required' => 'Harap pilih penulis yang tersedia',
+                    'in_list' => 'Harap pilih penulis yang tesedia',
                 ]
             ],
             'penerbit' => [
-                'rules' => 'required',
+                'rules' => 'required|is_not_unique[penerbit.id_penerbit]',
                 'errors' => [
-                    'required' => 'Kolom penerbit tidak boleh kosong'
+                    'required' => 'Harap pilih penerbit yang tersedia',
+                    'in_list' => 'Harap pilih penerbit yang tesedia',
+                ]
+            ],
+            'kategori' => [
+                'rules' => 'required|is_not_unique[kategori.id_kategori]',
+                'errors' => [
+                    'required' => 'Harap pilih kategori yang tersedia',
+                    'in_list' => 'Harap pilih kategori yang tesedia',
+                ]
+            ],
+            'rak' => [
+                'rules' => 'required|is_not_unique[rak.id_rak]',
+                'errors' => [
+                    'required' => 'Harap pilih rak yang tersedia',
+                    'in_list' => 'Harap pilih rak yang tesedia',
                 ]
             ],
             'tanggal_terbit' => [
@@ -96,12 +129,6 @@ class Buku extends BaseController
                     'numeric' => 'Harap isi kolom jumlah buku menggunakan angka',
                 ]
             ],
-            'sinopsis' => [
-                'rules' => 'required',
-                'errors' => [
-                    'required' => 'Kolom sinopsis tidak boleh kosong'
-                ]
-            ],
         ])) {
             return redirect()->back()->withInput()->with('error_tambah', true);
         }
@@ -115,46 +142,56 @@ class Buku extends BaseController
             $fileSampul->move('upload/buku/', $namasampul);
         }
 
-        $this->bukuModel->save([
+        $save = $this->buku->save([
             'id_buku' => uniqueID('BK', 'buku', 'id_buku'),
             'judul' => $this->request->getVar('judul'),
             'slug' =>  url_title($this->request->getVar('judul'), '-', true),
-            'penulis' => $this->request->getVar('penulis'),
-            'penerbit' => $this->request->getVar('penerbit'),
+            'id_penulis' => $this->request->getVar('penulis'),
+            'id_penerbit' => $this->request->getVar('penerbit'),
+            'id_kategori' => $this->request->getVar('kategori'),
+            'id_rak' => $this->request->getVar('rak'),
             'tanggal_terbit' => $this->request->getVar('tanggal_terbit'),
             'jumlah_buku' => $this->request->getVar('jumlah_buku'),
             'sampul' => $namasampul,
             'sinopsis' => $this->request->getVar('sinopsis'),
         ]);
 
-        session()->setFlashdata('pesan', "Data berhasil ditambahkan");
-
-        return redirect()->to("/admin/buku");
+        if ($save) {
+            return redirect()->to("/admin/buku")->with('pesan', "Data berhasil ditambahkan");
+        }
+        return redirect()->back()->with('error', "Data gagal ditambahkan, Coba muat ulang halaman");
     }
 
     public function hapus($id)
     {
-        $buku = $this->bukuModel->where('id_buku', $id)->first();
+        $buku = $this->buku->find($id);
+        if (!$buku) {
+            return redirect()->back()->with('error', 'Buku yang ingin dihapus tidak ada, Harap ulangi proses atau muat ulang website');
+        }
+        $buku = $this->buku->where('id_buku', $id)->first();
         if ($buku['sampul'] != "default.png") {
             unlink('upload/buku/' . $buku['sampul']);
         }
 
-        $this->bukuModel->delete($id);
-
-        session()->setFlashdata('pesan', "Data berhasil dihapus");
-
-        return redirect()->to('/admin/buku');
+        if ($this->buku->delete($id)) {
+            return redirect()->to('/admin/buku')->with('pesan', "Data berhasil dihapus");
+        }
+        return redirect()->back()->with('pesan', "Data gagal dihapus");
     }
 
     public function edit($slug)
     {
-        $buku = $this->bukuModel->ambilBuku($slug);
+        $buku = $this->buku->getDataBySlug($slug);
+        if (!$buku) {
+            return redirect()->back()->with('error', 'Buku yang ingin diubah tidak ada, Harap ulangi proses atau muat ulang website');
+        }
+
+        $buku = $this->buku->getDataBySlug($slug);
         $id = $buku['id_buku'];
         $ruleJudul = 'required';
         if ($buku['judul'] != $this->request->getVar('judul')) {
             $ruleJudul .= '|is_unique[buku.judul]';
         }
-
 
         if (!$this->validate([
             'judul' => [
@@ -165,15 +202,31 @@ class Buku extends BaseController
                 ]
             ],
             'penulis' => [
-                'rules' => 'required',
+                'rules' => 'required|is_not_unique[penulis.id_penulis]',
                 'errors' => [
-                    'required' => 'Kolom penulis tidak boleh kosong'
+                    'required' => 'Harap pilih penulis yang tersedia',
+                    'is_not_unique' => 'Harap pilih penulis yang tesedia',
                 ]
             ],
             'penerbit' => [
-                'rules' => 'required',
+                'rules' => 'required|is_not_unique[penerbit.id_penerbit]',
                 'errors' => [
-                    'required' => 'Kolom penerbit tidak boleh kosong'
+                    'required' => 'Harap pilih penerbit yang tersedia',
+                    'is_not_unique' => 'Harap pilih penerbit yang tesedia',
+                ]
+            ],
+            'kategori' => [
+                'rules' => 'required|is_not_unique[kategori.id_kategori]',
+                'errors' => [
+                    'required' => 'Harap pilih kategori yang tersedia',
+                    'is_not_unique' => 'Harap pilih kategori yang tesedia',
+                ]
+            ],
+            'rak' => [
+                'rules' => 'required|is_not_unique[rak.id_rak]',
+                'errors' => [
+                    'required' => 'Harap pilih rak yang tersedia',
+                    'is_not_unique' => 'Harap pilih rak yang tesedia',
                 ]
             ],
             'tanggal_terbit' => [
@@ -198,12 +251,6 @@ class Buku extends BaseController
                     'numeric' => 'Harap isi kolom jumlah dengan angka',
                 ]
             ],
-            'sinopsis' => [
-                'rules' => 'required',
-                'errors' => [
-                    'required' => 'Kolom sinopsis tidak boleh kosong'
-                ]
-            ],
         ])) {
             return redirect()->back()->withInput()->with('error_edit', true);
         }
@@ -219,20 +266,24 @@ class Buku extends BaseController
         }
         $slug = url_title($this->request->getVar('judul'), '-', true);
 
-        $this->bukuModel->save([
+        $save = $this->buku->save([
             'id_buku' => $id,
             'judul' => $this->request->getVar('judul'),
             'slug' =>  $slug,
-            'penulis' => $this->request->getVar('penulis'),
-            'penerbit' => $this->request->getVar('penerbit'),
+            'id_penulis' => $this->request->getVar('penulis'),
+            'id_penerbit' => $this->request->getVar('penerbit'),
+            'id_kategori' => $this->request->getVar('kategori'),
+            'id_rak' => $this->request->getVar('rak'),
             'tanggal_terbit' => $this->request->getVar('tanggal_terbit'),
             'sampul' => $namasampul,
             'jumlah_buku' => $this->request->getVar('jumlah_buku'),
             'sinopsis' => $this->request->getVar('sinopsis'),
         ]);
 
-        session()->setFlashdata('pesan', "Data berhasil diubah");
-
-        return redirect()->to("/admin/buku/" . $slug);
+        if ($save) {
+            return redirect()->to("/admin/buku/" . $slug)->with('pesan', "Data berhasil diubah");
+        }
+        
+        return redirect()->to("/admin/buku/" . $slug)->with('pesan', "Data gagal diubah");
     }
 }
